Changed the timing of event dispatch execution in fire_event().
The event handler calls were always delayed one tick once the event was fired. In some cases, we wanted to get the events fired in the order of fire_event() as
possible. When the fire_event() happens in the EventMachine's thread, it became to dispatch the event handlers immediatly. The delay occurs only
when the fire_event() is called from the other worker threads.