Added lock queue for actions.
This feature helps to solve the conflictions between actions. The story behind is that multiple actions may collapse the internal data. When the new job/action is kicked off, it may take a while to process. In the mean time, someone starts another new job/action. It potentially involves internal data structure collapse since two or more different jobs run at same time which treat same data.

To avoid the confliction, it is important to defer the later job to run until the running job finish. The lock queue tracks the running jobs and waiting jobs and also notifies to the waiting jobs when it's ready to start.