Modifiy to handle multiple queues for each thread in LockQueue.
The last implementation just used single queue for getting notification on quit(). It was not right way to get
the notification since the relation between notifier and receiver is M:N. This change adds a queue list @queue_by_thread
and a mutex lock to the class.

Some barrier codes added to make it more thread safe.

LockQueue#inspect() had a bug when it shows empty lock object. This issue is also fixed.