Changed the registration packet to include agent attributes.
Prior to this change, the Registration packet sent only limited information. The master runs
some triggers when the packet arrived. But some of them need the detail information about
agent. The Ping packet tells such details. There is arraival time gap between Registration and Ping.
In worst case, the triggers face crash due to the lack of agent details when the Registration packet came but not Ping yet.

The solution given by this patch is adding the agent detail information to the Registration packet as well as Ping does.