Improved concurrency on propagating service instances.
Starting new VM instance was called as a part of sequence on propagation. It is a syncrhonous process to hold a worker thread till
it is confirmed to be online. Create a new action for the VM instance start up and modify the steps of agent allocation part to get
whole process worked concurrently.