Add PropagateResource server command and action.
Propagate action is splitted to two different actions: propagate_service and propagate_resource.
propagate_service uses the service instance ID of existing one as the reference object.
propagate_resource accepts the resource class name for the reference and Host ID already registered in the cluster. Those two are mandated for the resource propagation.