Rename RuleEngine to ActionManager.
ActionManager class takes charge of concurrency control for Actions. RuleEngine used to be managing them but
it was associated with ServiceCluster instances separatly. This is not reasonable design that the actions also associated
to the ServiceClusters.
ActionManager is one of master's manager module. It handles all the running actions accross the master process.