Add ServiceClusterPool model.
The service clusters were lost after the master process is restarted even though the database still
stores the details about the clusters. ServiceClusterPool is the model to provide single point resource
for the cluster ID list. Retrieve the IDs from this model when it is needed to iterate the cluster data
imported already.