Update class methods in Subcommand module to attribute subcommand meta data.
The class setter methods were defined to use the equal operator to set values:
  module Subcommand
    def self.command_name=(name)
    .....
    end
  end

But the above did not work in the example code below:
  class A
    include Subcommand
    command_name = 'a_subcommand'
  end

The left item before the equal operator is understood as an new variable by the interpreter so that
the setter method was never called.

In this fix, the setter methods were changed to accept the assigned values as method argument.
   def self.command_name(name)
   ....
   end