Change method interface for agent monitors.
The old interface was using register/unregister method calls for each monitor setup:

1. Master remotely calls Agent.monitors.register(svc_id, *args) to start the monitoring.
2. Agent starts the monitor thread with the given parameters.
3. Master remotely calls Agent.monitors.unregister(svc_id) to stop.
4. Agent stops the monitor thread.

This easily causes discrepancies between what the master wanted to monitor and the agent
is monitoring when the failure event occurs in one side (even both).
To avoid this, the agent should accept the monitoring configuration in one batch:

1. Master sends a set of conf hash for what it wants to watch on the Agent.
2. The Agent stops the running monitors then restarts the monitors based on the new conf.

This commit just includes the changes for the Agent end. The Master end also has to be changed.