Change the agent registration process.
For the VM attribute, the agent send the data periodically through the registration and the ping packets. This design
is not the best way because the agent has responsibility on fetching the these attributes even if it is not capable of.
i.e. EC2 allows each VM to fetch using the internal metadata server but other clouds do not.

In the new procedure, the agent sends its unique ID then the master collects the detail VM information by requesting
arbitrary cloud API server with the ID. The ID format is not determined yet though.