Improve service instance status/lifecycle management.
ServiceInstance#status was used to represent each service how it is working. There were
two players who can update this variable: Action and Agent. This caused that the single variable was assignable
two different types of information: live status and lifecycle status even though they are different.

This change adds ServiceInstance#monitor_status which is to save the online/offline status of the service
notified by the agent. ServiceInstance#status is also changed to show the lifecycle status of the service.
(such like migrating, starting or running)