Change the initialization order of the master process.
There was an issue that the master process creates its database file with wrong owner. I am not sure
why it happened at this timing but decided to change the setup order in the master.

The master process was displaying a few log messages to STDOUT before getting the daemon mode started. This
change fixed to get all the messages to log file.