Fix the critical bug which stucks the running thread up.
EM.barrier() and StatusDB.barrier() are used for gather the arbitrary block
to the worker thread. barrier() is designed to stop the caller thread while the passed block
is being processed in the worker thread. To do that, Queue object is utilized to stop the thread.
The root problem was that the Queue object is storted on to the shared variable, class/object variable
in this case, although the variable scope had to be method local.

The fix is just to change the variable scope.