Wrap with StatusDB.barrier for all StatusDB's DB access method calls.
This change makes that the all StatusDB primitive DB accesses, find/save/delete, happen
in the single worker thread sequentially. This helps to improve the access to the sqlite
database since that is using the database lock.

For other database drivers, this feature may not be appropriate for concurrency. It may need
a flag to disable to push into one worker thread.